// src/app/models/entities.ts
export interface Role { 
    Id?: number;
    OpUserId?: number;
    OpDate?: Date;
    OpIp?: string;
    Name?: string;
    IsAdmin?: boolean;
}

export interface RoleAction { 
    Id?: number;
    OpUserId?: number;
    OpDate?: Date;
    OpIp?: string;
    RoleId?: number;
    ActionId?: number;
    Enabled?: boolean;
}

export interface RoleMenu { 
    Id?: number;
    OpUserId?: number;
    OpDate?: Date;
    OpIp?: string;
    RoleId?: number;
    MenuId?: number;
    HasAccess?: boolean;
}

export interface SaveRoleActionsModel { 
    RoleName?: string;
    Type?: number;
    Data?: TreeNode[];
}
// ...
export interface TreeNode { 
    label?: string;
    data?: any;
    icon?: string;
    expandedIcon?: string;
    collapsedIcon?: string;
    children?: TreeNode[];
    leaf?: boolean;
    expanded?: boolean;
    type?: string;
    parent?: TreeNode;
    partialSelected?: boolean;
    checked?: boolean;
}